+++ saos-persistence/src/main/java/pl/edu/icm/saos/persistence/repository/ScJudgmentRepository.java
@@ -8,6 +8,6 @@ import pl.edu.icm.saos.persistence.model.SupremeCourtJudgment;
  * @author Łukasz Dumiszewski
  */
 
-public interface ScJudgmentRepository extends JudgmentCommonRepository<SupremeCourtJudgment>, JpaRepository<SupremeCourtJudgment, Integer> {
+public interface ScJudgmentRepository extends JudgmentCommonRepository<SupremeCourtJudgment>, JpaRepository<SupremeCourtJudgment, Long> {
 
 }
