+++ saos-persistence/src/main/java/pl/edu/icm/saos/persistence/search/implementor/CommonCourtJpqlSearchImplementator.java
@@ -1,17 +1,19 @@
 package pl.edu.icm.saos.persistence.search.implementor;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import pl.edu.icm.saos.persistence.model.CommonCourt;
-import pl.edu.icm.saos.persistence.search.dto.CommonCourtSearchFilter;
-import pl.edu.icm.saos.persistence.search.result.SearchResult;
-
-import javax.persistence.EntityManager;
-import javax.persistence.Query;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
+
+import javax.persistence.EntityManager;
+import javax.persistence.Query;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import pl.edu.icm.saos.persistence.model.CommonCourt;
+import pl.edu.icm.saos.persistence.search.dto.CommonCourtSearchFilter;
+import pl.edu.icm.saos.persistence.search.result.SearchResult;
 
 /**
  * @author pavtel
@@ -42,7 +44,7 @@ public class CommonCourtJpqlSearchImplementator extends AbstractJpqlSearchImplem
 
     @Override
     protected void processResult(SearchResult<CommonCourt> searchResult, CommonCourtSearchFilter searchFilter) {
-        List<Integer> courtsIds = extractCommonCourtIds(searchResult);
+        List<Long> courtsIds = extractCommonCourtIds(searchResult);
         if(courtsIds.isEmpty()){
             return;
         }
@@ -55,17 +57,17 @@ public class CommonCourtJpqlSearchImplementator extends AbstractJpqlSearchImplem
 
     //******** PRIVATE *************
 
-    private List<Integer> extractCommonCourtIds(SearchResult<CommonCourt> searchResult) {
+    private List<Long> extractCommonCourtIds(SearchResult<CommonCourt> searchResult) {
         return searchResult.getResultRecords().stream().map(result->result.getId()).collect(Collectors.toList());
     }
 
-    private void initializeCommonCourtDivisions(List<Integer> courtsIds){
+    private void initializeCommonCourtDivisions(List<Long> courtsIds){
         setIdsParameterAndExecuteQuery(" select commonCourt from " + CommonCourt.class.getName() + " commonCourt left join fetch commonCourt.divisions_ division where commonCourt.id in (:ids) ",
                 courtsIds);
     }
 
 
-    private void setIdsParameterAndExecuteQuery(String query, List<Integer> ids){
+    private void setIdsParameterAndExecuteQuery(String query, List<Long> ids){
         Query queryObject = entityManager.createQuery(query);
         queryObject.setParameter("ids", ids);
